refactor network types into module
