+++ cmd/ethereum_rust/ethereum_rust.rs
@@ -2,6 +2,7 @@ use bytes::Bytes;
 use ethereum_rust_chain::add_block;
 use ethereum_rust_core::types::{Block, Genesis};
 use ethereum_rust_net::bootnode::BootNode;
+use ethereum_rust_net::types::Node;
 use ethereum_rust_storage::{EngineType, Store};
 use std::{
     fs::File,
